requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
python-dotenv
passlib[bcrypt]
bcrypt<5
python-jose[cryptography]
pydantic[email]
sqlalchemy
asyncpg
alembic
python-multipart
pytest
pytest-asyncio
httpx
psycopg[binary]
google-auth
requests
